pyproject.toml
[tool.poetry]
name = "base-date-vacancy"
version = "0.1.0"
description = ""
authors = ["Meira Beloritski"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
flake8 = "^7.1.1"
pytest = "^8.3.2"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
flake8 = "^7.1.1"
mypy = "^1.11.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.black]
line-length = 119
target-version = ['py312']
skip-string-normalization = false
exclude = """
/(
    \\.git
  | \\.mypy_cache
  | \\.tox
  | \\.venv
  | _build
  | buck-out
  | build
  | dist
)/
"""